move format data
